+++ R/gaps.R
@@ -10,14 +10,14 @@ globalVariables(c(".", ".gaps"))
 #' existing `NA` untouched.
 #' * empty: filled with default `NA`.
 #' * filled by values or functions.
-#' @param .full `FALSE` to insert `NA` for each series within its own period. `TRUE`
-#' to fill `NA` over the entire time span of the data (a.k.a. fully balanced panel).
+#' @param .full `FALSE` inserts `NA` for each keyed unit within its own period. `TRUE`
+#' fills `NA` over the entire time span of the data (a.k.a. fully balanced panel).
+#' Other options: `start()` or `end()` at the same time.
 #'
 #' @family implicit gaps handling
 #' @seealso [tidyr::fill], [tidyr::replace_na] for handling missing values `NA`.
 #' @export
 #' @examples
-#' library(dplyr)
 #' harvest <- tsibble(
 #'   year = c(2010, 2011, 2013, 2011, 2012, 2014),
 #'   fruit = rep(c("kiwi", "cherry"), each = 3),
@@ -27,6 +27,8 @@ globalVariables(c(".", ".gaps"))
 #'
 #' # gaps as default `NA`
 #' fill_gaps(harvest, .full = TRUE)
+#' fill_gaps(harvest, .full = start())
+#' fill_gaps(harvest, .full = end())
 #' full_harvest <- fill_gaps(harvest, .full = FALSE)
 #' full_harvest
 #'
@@ -76,7 +78,7 @@ fill_gaps.tbl_ts <- function(.data, ..., .full = FALSE) {
   nrows <- vec_size(.data)
   if (nrows == 0L || nrows == 1L) return(.data)
 
-  gap_data <- scan_gaps(.data, .full = .full)
+  gap_data <- scan_gaps(.data, .full = !!enquo(.full))
   lst_exprs <- enquos(..., .named = TRUE)
   if (vec_size(gap_data) == 0 && !is_empty(lst_exprs)) return(.data)
 
@@ -110,6 +112,7 @@ scan_gaps <- function(.data, .full = FALSE) {
 
 #' @export
 scan_gaps.tbl_ts <- function(.data, .full = FALSE) {
+  .full <- quo_get_expr(enquo(.full))
   int <- interval(.data)
   idx <- index(.data)
   idx_chr <- as_string(idx)
@@ -119,14 +122,30 @@ scan_gaps.tbl_ts <- function(.data, .full = FALSE) {
 
   key <- key(.data)
   keyed_tbl <- new_grouped_df(.data, groups = key_data(.data))
-  if (.full) {
+  if (is_true(.full)) {
     idx_full <- seq_generator(keyed_tbl[[idx_chr]], int)
     sum_data <-
       summarise(keyed_tbl, !!idx_chr := list2(tibble(!!idx_chr := idx_full)))
-  } else {
+  } else if (is_false(.full)) {
     sum_data <- summarise(keyed_tbl,
       !!idx_chr := list2(tibble(!!idx_chr := seq_generator(!!idx, int)))
     )
+  } else if (.full == sym("start()")) {
+    start <- min(keyed_tbl[[idx_chr]])
+    sum_data <- summarise(keyed_tbl,
+      !!idx_chr := list2(tibble(
+        !!idx_chr := seq_generator(c(start, max(!!idx)), int)
+      ))
+    )
+  } else if (.full == sym("end()")) {
+    end <- max(keyed_tbl[[idx_chr]])
+    sum_data <- summarise(keyed_tbl,
+      !!idx_chr := list2(tibble(
+        !!idx_chr := seq_generator(c(min(!!idx), end), int)
+      ))
+    )
+  } else {
+    abort_invalid_full_arg()
   }
   ref_data <- unwrap(sum_data, !!idx)
   if (vec_size(ref_data) == vec_size(.data)) {
@@ -139,9 +158,7 @@ scan_gaps.tbl_ts <- function(.data, .full = FALSE) {
 
 #' Count implicit gaps
 #'
-#' @param .data A `tbl_ts`.
-#' @param .full `FALSE` to find gaps for each series within its own period.
-#' `TRUE` to find gaps over the entire time span of the data.
+#' @inheritParams fill_gaps
 #' @param .name Strings to name new columns.
 #'
 #' @family implicit gaps handling
@@ -170,7 +187,7 @@ count_gaps <- function(.data, .full = FALSE, .name = c(".from", ".to", ".n")) {
   int <- interval(.data)
   idx <- index(.data)
 
-  gap_data <- scan_gaps(.data, .full = .full)
+  gap_data <- scan_gaps(.data, .full = !!enquo(.full))
   if (vec_size(gap_data) == 0L) {
     data_key <- .data[0L, key_vars(.data)]
     idx_vals <- .data[[as_string(idx)]][0L]
@@ -183,9 +200,7 @@ count_gaps <- function(.data, .full = FALSE, .name = c(".from", ".to", ".n")) {
   lst_out <- summarise(grped_tbl,
     !!".gaps" := list2(tbl_gaps(!!idx, idx_full, .name = .name)))
 
-  idx_type <- class(lst_out[[".gaps"]][[1]][[.name[1]]])
   out <- unwrap(lst_out, .gaps)
-  class(out[[.name[1]]]) <- class(out[[.name[2]]]) <- idx_type
   tibble(!!!out)
 }
 
@@ -204,6 +219,8 @@ count_gaps <- function(.data, .full = FALSE, .name = c(".from", ".to", ".n")) {
 #' )
 #' has_gaps(harvest)
 #' has_gaps(harvest, .full = TRUE)
+#' has_gaps(harvest, .full = start())
+#' has_gaps(harvest, .full = end())
 has_gaps <- function(.data, .full = FALSE, .name = ".gaps") {
   stopifnot(has_length(.name, 1))
   if (!is_regular(.data) || vec_size(.data) == 0L) {
@@ -211,17 +228,31 @@ has_gaps <- function(.data, .full = FALSE, .name = ".gaps") {
     return(tibble(!!!key_data, !!.name := FALSE))
   }
 
+  .full <- quo_get_expr(enquo(.full))
   int <- interval(.data)
   idx <- index(.data)
+  idx_chr <- as_string(idx)
   grped_tbl <- new_grouped_df(.data, groups = key_data(.data))
-  if (.full) {
-    idx_full <- seq_generator(.data[[as_string(idx)]], int)
+  if (is_true(.full)) {
+    idx_full <- seq_generator(.data[[idx_chr]], int)
     res <- summarise(grped_tbl,
       !!.name := (length(idx_full) - length(!!idx)) > 0)
-  } else {
+  } else if (is_false(.full)) {
     res <- summarise(grped_tbl,
       !!.name := (length(seq_generator(!!idx, int)) - length(!!idx)) > 0
     )
+  } else if (.full == sym("start()")) {
+    start <- min(.data[[idx_chr]])
+    res <- summarise(grped_tbl,
+      !!.name := (length(seq_generator(c(start, max(!!idx)), int)) - length(!!idx)) > 0
+    )
+  } else if (.full == sym("end()")) {
+    end <- max(.data[[idx_chr]])
+    res <- summarise(grped_tbl,
+      !!.name := (length(seq_generator(c(min(!!idx), end), int)) - length(!!idx)) > 0
+    )
+  } else {
+    abort_invalid_full_arg()
   }
   tibble(!!!res)
 }
@@ -235,6 +266,9 @@ tbl_gaps <- function(x, y, .name = c(".from", ".to", ".n")) {
       "`length(x)` (%d) must not be greater than `length(y)` (%d).",
       len_x, len_y
     ))
+  }
+  if (len_x == 0) {
+    return(tibble(!!.name[1] := y, !!.name[2] := y, !!.name[3] := integer()))
   }
   gap_vec <- logical(length = len_y)
   gap_vec[vec_match(x, y)] <- TRUE
@@ -292,4 +326,8 @@ unwrap <- function(.data, .col) {
   )
   res <- vec_slice(res, row_indices)[setdiff(names(.data), lst_col)]
   vec_cbind(res, vec_rbind(!!!.data[[lst_col]]))
+}
+
+abort_invalid_full_arg <- function() {
+  abort("`.full` only accepts `TRUE`, `FALSE`, `start()`, or `end()`.")
 }